Remove base manager abstraction
