+++ examples/crewai/crewai_with_arcade_tool.py
@@ -10,11 +10,14 @@ The example assumes the following:
 
 """
 
+from arcadepy import Arcade
 from crewai import Agent, Crew, Task
 from crewai.llm import LLM
-from crewai_arcade.manager import CrewAIToolManager
+from crewai_arcade.manager import ArcadeToolManager
 
-manager = CrewAIToolManager(user_id="user@example.com")
+manager = ArcadeToolManager(
+    user_id="user@example.com", client=Arcade(base_url="http://localhost:9099")
+)
 tools = manager.get_tools(tools=["Google.ListEmails"])
 
 crew_agent = Agent(
