+++ examples/crewai/requirements.txt
@@ -1,2 +1,1 @@
 crewai-arcade>=0.1.0
-arcade-google>=0.2.0
